ECWID-145233 Change IDs type from int to long (bigint) in node: Category entity, id field.
Modified the corresponding classes to make the category id type to Long.

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/cart/result/CalculateOrderDetailsResult.kt
@@ -126,14 +126,14 @@ data class CalculateOrderDetailsResult(
 
 	data class DiscountCouponCatalogLimit(
 		val products: List<Int>? = null,
-		val categories: List<Int>? = null
+		val categories: List<Long>? = null
 	)
 
 	data class OrderItem(
 		val id: Int? = null,
 
 		val productId: Int? = null,
-		val categoryId: Int? = null,
+		val categoryId: Long? = null,
 
 		val price: Double? = null,
 		val productPrice: Double? = null,
